+++ Changelog.md
@@ -12,6 +12,7 @@
   * todo: document this, add non-generic version, add path based versions
 * added `ReadableWritableDataFile.ResetValueToDefault()`
 * added `ReadableWritableDataFile.ResetToDefaultData()`
+* added `BaseTypes.TryParseBaseType`
 * fixed default data path on Mac OS (Unity version)
 * added `ReadableDataFile.Identifier`
   * when creating a `MemoryDataFile`, you can now supply an identifier